+++ code/src/Plexdata.CsvParser.NET.Tests/Helpers/ProcessHelperTests.cs
@@ -24,8 +24,10 @@
 
 using NUnit.Framework;
 using Plexdata.CsvParser.Helpers;
+using Plexdata.CsvParser.Processors;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Text;
 
 namespace Plexdata.CsvParser.Tests.Helpers
@@ -43,6 +45,83 @@ namespace Plexdata.CsvParser.Tests.Helpers
             List<String> actual = ProcessHelper.SplitIntoCells(item.Value, item.Separator);
 
             Assert.That(String.Join(String.Empty, actual), Is.EqualTo(String.Join(String.Empty, item.Expected)));
+        }
+
+        [Test]
+        [TestCase("Quoting on", true)]
+        [TestCase(42, false)]
+        public void ConvertToString_QuotingIsTrueForStrings_ResultAsExpected(Object value, Boolean expected)
+        {
+            CultureInfo culture = CultureInfo.CurrentUICulture;
+            CsvMappings mapping = new CsvMappings();
+
+            ProcessHelper.ConvertToString(value, culture, mapping, out Boolean actual);
+
+            Assert.That(actual, Is.EqualTo(expected));
+        }
+
+        [Test]
+        [TestCase(null, ':', false, ":")]
+        [TestCase(null, ':', true, "\"\":")]
+        [TestCase("\"", ':', false, "\"\"\"\":")]
+        [TestCase("\"", ':', true, "\"\"\"\":")]
+        [TestCase(":", ':', false, "\":\":")]
+        [TestCase(":", ':', true, "\":\":")]
+        [TestCase(":\"", ':', false, "\":\"\"\":")]
+        [TestCase(":\"", ':', true, "\":\"\"\":")]
+        [TestCase("\r", ':', false, "\"\r\":")]
+        [TestCase("\r", ':', true, "\"\r\":")]
+        [TestCase("\r\"", ':', false, "\"\r\"\"\":")]
+        [TestCase("\r\"", ':', true, "\"\r\"\"\":")]
+        [TestCase("\n", ':', false, "\"\n\":")]
+        [TestCase("\n", ':', true, "\"\n\":")]
+        [TestCase("\n\"", ':', false, "\"\n\"\"\":")]
+        [TestCase("\n\"", ':', true, "\"\n\"\"\":")]
+        [TestCase("\r\n", ':', false, "\"\r\n\":")]
+        [TestCase("\r\n", ':', true, "\"\r\n\":")]
+        [TestCase("\r\n\"", ':', false, "\"\r\n\"\"\":")]
+        [TestCase("\r\n\"", ':', true, "\"\r\n\"\"\":")]
+        [TestCase("\n:", ':', false, "\"\n:\":")]
+        [TestCase("\n:", ':', true, "\"\n:\":")]
+        [TestCase("\n\":", ':', false, "\"\n\"\":\":")]
+        [TestCase("\n\":", ':', true, "\"\n\"\":\":")]
+        [TestCase("\r\n:", ':', false, "\"\r\n:\":")]
+        [TestCase("\r\n:", ':', true, "\"\r\n:\":")]
+        [TestCase("\r\n\":", ':', false, "\"\r\n\"\":\":")]
+        [TestCase("\r\n\":", ':', true, "\"\r\n\"\":\":")]
+        [TestCase("\"quote at front", ':', false, "\"\"\"quote at front\":")]
+        [TestCase("\"quote at front", ':', true, "\"\"\"quote at front\":")]
+        [TestCase("quote at end\"", ':', false, "\"quote at end\"\"\":")]
+        [TestCase("quote at end\"", ':', true, "\"quote at end\"\"\":")]
+        [TestCase("quote in\"middle", ':', false, "\"quote in\"\"middle\":")]
+        [TestCase("quote in\"middle", ':', true, "\"quote in\"\"middle\":")]
+        [TestCase("nothing to quote", ':', false, "nothing to quote:")]
+        [TestCase("nothing to quote", ':', true, "\"nothing to quote\":")]
+        public void ConvertToOutput_ValueAutoQuoting_ResultAsExpected(String value, Char separator, Boolean quoting, String expected)
+        {
+            Assert.That(ProcessHelper.ConvertToOutput(value, separator, quoting), Is.EqualTo(expected));
+        }
+
+        [Test]
+        public void FixupOutput_StringBuilderIsNull_ResultIsNull()
+        {
+            Assert.That(ProcessHelper.FixupOutput(null, '#'), Is.Null);
+        }
+
+        [Test]
+        public void FixupOutput_StringBuilderIsEmpty_ResultIsEmpty()
+        {
+            Assert.That(ProcessHelper.FixupOutput(new StringBuilder(), '#').Length, Is.EqualTo(0));
+        }
+
+        [Test]
+        [TestCase("some data", '#', "some data")]
+        [TestCase("some data#", '#', "some data")]
+        [TestCase("some#data", '#', "some#data")]
+        [TestCase("some#data#", '#', "some#data")]
+        public void FixupOutput_StringBuilderWithSomeData_ResultAsExpected(String value, Char separator, String expected)
+        {
+            Assert.That(ProcessHelper.FixupOutput(new StringBuilder(value), separator).ToString(), Is.EqualTo(expected));
         }
 
         #region Test helper classes