+++ code/src/Plexdata.CsvParser.Tester/Program.cs
@@ -195,37 +195,77 @@ namespace Plexdata.CsvParser.Tester
         /* Read CSV file
         static void Main(String[] args)
         {
-            // Source file could contain this content:
-            // Name;               Notes
-            // "Marley, Bob";      "Jamaican singer-songwriter"
-            // "Monroe, Marilyn";  "American actress";          "model and singer"
-            // "Snipes, Wesley";   "American actor";            "director, film producer"; "martial artist"
-            // "Hurley, Elizabeth" 
+            try
+            {
+                // Source file could contain this content:
+                // Name;               Notes
+                // "Marley, Bob";      "Jamaican singer-songwriter"
+                // "Monroe, Marilyn";  "American actress";          "model and singer"
+                // "Snipes, Wesley";   "American actor";            "director, film producer"; "martial artist"
+                // "Hurley, Elizabeth" 
 
-            CsvSettings settings = new CsvSettings() { Heading = true, Separator = ColumnSeparators.SemicolonSeparator };
-            CsvContainer container = CsvReader.Read(filename, settings);
+                CsvSettings settings = new CsvSettings() { Heading = true, Separator = ColumnSeparators.SemicolonSeparator };
+                CsvContainer container = CsvReader.Read(filename, settings);
 
-            String col0row1 = container.GetValue<String>(0, 1) as String; // Marley, Bob
-            String col0row2 = container.GetValue<String>(0, 2) as String; // Monroe, Marilyn
-            String col0row3 = container.GetValue<String>(0, 3) as String; // Snipes, Wesley
-            String col0row4 = container.GetValue<String>(0, 4) as String; // Hurley, Elizabeth
+                String col0row1 = container.GetValue<String>(0, 1) as String; // Marley, Bob
+                String col0row2 = container.GetValue<String>(0, 2) as String; // Monroe, Marilyn
+                String col0row3 = container.GetValue<String>(0, 3) as String; // Snipes, Wesley
+                String col0row4 = container.GetValue<String>(0, 4) as String; // Hurley, Elizabeth
 
-            String col1row1 = container.GetValue<String>(1, 1) as String; // Jamaican singer-songwriter
-            String col1row2 = container.GetValue<String>(1, 2) as String; // American actress
-            String col1row3 = container.GetValue<String>(1, 3) as String; // American actor
-            String col1row4 = container.GetValue<String>(1, 4) as String; // null
+                String col1row1 = container.GetValue<String>(1, 1) as String; // Jamaican singer-songwriter
+                String col1row2 = container.GetValue<String>(1, 2) as String; // American actress
+                String col1row3 = container.GetValue<String>(1, 3) as String; // American actor
+                String col1row4 = container.GetValue<String>(1, 4) as String; // null
 
-            String col2row1 = container.GetValue<String>(2, 1) as String; // null
-            String col2row2 = container.GetValue<String>(2, 2) as String; // model and singer
-            String col2row3 = container.GetValue<String>(2, 3) as String; // director, film producer
-            String col2row4 = container.GetValue<String>(2, 4) as String; // null
+                String col2row1 = container.GetValue<String>(2, 1) as String; // null
+                String col2row2 = container.GetValue<String>(2, 2) as String; // model and singer
+                String col2row3 = container.GetValue<String>(2, 3) as String; // director, film producer
+                String col2row4 = container.GetValue<String>(2, 4) as String; // null
 
-            String col3row1 = container.GetValue<String>(3, 1) as String; // null
-            String col3row2 = container.GetValue<String>(3, 2) as String; // null
-            String col3row3 = container.GetValue<String>(3, 3) as String; // martial artist
-            String col3row4 = container.GetValue<String>(3, 4) as String; // null
+                String col3row1 = container.GetValue<String>(3, 1) as String; // null
+                String col3row2 = container.GetValue<String>(3, 2) as String; // null
+                String col3row3 = container.GetValue<String>(3, 3) as String; // martial artist
+                String col3row4 = container.GetValue<String>(3, 4) as String; // null
 
-            Console.ReadKey();
+                Console.ReadKey();
+            }
+            catch (Exception exception)
+            {
+                Console.WriteLine(exception);
+            }
+        }
+        */
+
+        /* Write CSV file
+        static void Main(String[] args)
+        {
+            try
+            {
+                // Output file would contain this content:
+                // Name;Notes
+                // "Marley, Bob";"Jamaican singer-songwriter"
+                // "Monroe, Marilyn";"American actress";"model and singer"
+                // "Snipes, Wesley";"American actor";"director, film producer";"martial artist"
+                // "Hurley, Elizabeth" 
+
+                List<List<Object>> content = new List<List<Object>>
+                {
+                    new List<Object> { "Name", "Notes" },
+                    new List<Object> { "Marley, Bob", "Jamaican singer-songwriter" },
+                    new List<Object> { "Monroe, Marilyn", "American actress", "model and singer" },
+                    new List<Object> { "Snipes, Wesley", "American actor", "director, film producer", "martial artist" },
+                    new List<Object> { "Hurley, Elizabeth" }
+                };
+
+                CsvSettings settings = new CsvSettings() { Heading = true, Textual = true, Separator = ColumnSeparators.SemicolonSeparator };
+                CsvWriter.Write(content, filename, settings);
+
+                Console.ReadKey();
+            }
+            catch (Exception exception)
+            {
+                Console.WriteLine(exception);
+            }
         }
         */